% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ctmle.R
\name{print.ctmle}
\alias{print.ctmle}
\title{print a ctmle object}
\usage{
\method{print}{ctmle}(x, ...)
}
\arguments{
\item{x}{a ctmle object}

\item{...}{other parameter}
}
\description{
print a ctmle object
}
