% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ctmle.R
\name{summary.ctmle}
\alias{summary.ctmle}
\title{Summarise a ctmle object}
\usage{
\method{summary}{ctmle}(object, ...)
}
\arguments{
\item{object}{a ctmle object}

\item{...}{other parameter}
}
\description{
Summarise a ctmle object
}
