% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ctmle.R
\name{print.summary.ctmle}
\alias{print.summary.ctmle}
\title{print the summary of a ctmle object}
\usage{
\method{print}{summary.ctmle}(x, ...)
}
\arguments{
\item{x}{a summary.ctmle object}

\item{...}{other parameter}
}
\description{
print the summary of a ctmle object
}
